Remove ps(1) support from Illumos.  ps(1) args aren't wide enough to be useful.
Even with a shorter path, half of WAL filename being replayed would be missing.

```
$ /usr/bin/ps -o args -p `pgrep -P 56264 | xargs echo | sed -e 's# #,#g'`
COMMAND
/opt/local/bin/postgres -D /var/pgsql/go/src/github.com/joyent/pg_prefaulter/.p
/opt/local/bin/postgres -D /var/pgsql/go/src/github.com/joyent/pg_prefaulter/.p
/opt/local/bin/postgres -D /var/pgsql/go/src/github.com/joyent/pg_prefaulter/.p
/opt/local/bin/postgres -D /var/pgsql/go/src/github.com/joyent/pg_prefaulter/.p
/opt/local/bin/postgres -D /var/pgsql/go/src/github.com/joyent/pg_prefaulter/.p
/opt/local/bin/postgres -D /var/pgsql/go/src/github.com/joyent/pg_prefaulter/.p
```
